+++ changelog.md
@@ -51,6 +51,7 @@ Codion Change Log
 - EntityTablePanel.Config.conditionPanelFactory() renamed conditionPanel().
 - EntityEditPanel.InsertCommand, UpdateCommand and DeleteCommand.Builder.execute() added.
 - EntityTableExportModel bug fixed, now reloads configuration when files are saved or refreshed.
+- EntityTableExportPanel bug fixed, unselected foreign key attributes with selected children can now be moved.
 
 ## 0.18.60
 ### is.codion.common.reactive